jwt登录接口完成

+++ im_auth/auth_api/internal/config/config.go
@@ -9,6 +9,6 @@ type Config struct {
 	}
 	Auth struct {
 		AccessSecret string
-		AccessExpire int64
+		AccessExpire int
 	}
 }
